Legalize comm path from global to asram
