+++ testing/python/transform/test_tilelang_transform_sunmmio_global_layout_utils.py
@@ -104,7 +104,7 @@ def test_global_buffer_layout_populated_for_sunmmio():
     with tvm.target.Target(target):
         mod = tvm.tir.transform.BindTarget(target)(mod)
         mod = tl.transform.InferSramScope()(mod)
-        mod = tl.transform.LegalizeSunmmioCopyPath()(mod)
+        mod = tl.transform.LegalizeSunmmioDataPath()(mod)
         mod = tl.transform.LayoutReducer()(mod)
         mod = tl.transform.SunmmioLayoutInference()(mod)
         mod = tl.transform.LowerTileOp()(mod)
@@ -144,7 +144,7 @@ def test_global_buffer_layout_not_populated_for_cuda():
     with tvm.target.Target(target):
         mod = tvm.tir.transform.BindTarget(target)(mod)
         mod = tl.transform.InferSramScope()(mod)
-        mod = tl.transform.LegalizeSunmmioCopyPath()(mod)
+        mod = tl.transform.LegalizeSunmmioDataPath()(mod)
         mod = tl.transform.LayoutInference()(mod)
         CollectLayoutMap()(mod)
 
@@ -189,7 +189,7 @@ def test_row_major_global_layout_values():
     with tvm.target.Target(target):
         mod = tvm.tir.transform.BindTarget(target)(mod)
         mod = tl.transform.InferSramScope()(mod)
-        mod = tl.transform.LegalizeSunmmioCopyPath()(mod)
+        mod = tl.transform.LegalizeSunmmioDataPath()(mod)
         mod = tl.transform.LayoutReducer()(mod)
         mod = tl.transform.SunmmioLayoutInference()(mod)
         mod = tl.transform.LowerTileOp()(mod)
@@ -248,7 +248,7 @@ def test_dynamic_shape_global_buffer_layout():
     with tvm.target.Target(target):
         mod = tvm.tir.transform.BindTarget(target)(mod)
         mod = tl.transform.InferSramScope()(mod)
-        mod = tl.transform.LegalizeSunmmioCopyPath()(mod)
+        mod = tl.transform.LegalizeSunmmioDataPath()(mod)
         mod = tl.transform.LayoutReducer()(mod)
         mod = tl.transform.SunmmioLayoutInference()(mod)
         mod = tl.transform.LowerTileOp()(mod)
