+++ testing/python/transform/test_tilelang_transform_sunmmio_threadless_lower_and_legalize.py
@@ -176,8 +176,8 @@ def run_lower_and_legalize_cascade(mod, target):
     mod = tl_transform.InferSramScope()(mod)
     assert_threadless_invariants(mod, "InferSramScope")
 
-    mod = tl_transform.LegalizeSunmmioCopyPath()(mod)
-    assert_threadless_invariants(mod, "LegalizeSunmmioCopyPath")
+    mod = tl_transform.LegalizeSunmmioDataPath()(mod)
+    assert_threadless_invariants(mod, "LegalizeSunmmioDataPath")
 
     mod = tl_transform.LayoutReducer()(mod)
     assert_threadless_invariants(mod, "LayoutReducer")
