+++ tilelang/engine/phase.py
@@ -174,8 +174,8 @@ def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
     mod = tilelang.transform.Simplify()(mod)
     # Infer shared memory SRAM scope
     mod = tilelang.transform.InferSramScope()(mod)
-    # Split Sunmmio global->asram copies before layout inference and tile-op lowering
-    mod = tilelang.transform.LegalizeSunmmioCopyPath()(mod)
+    # Stage unsupported Sunmmio global->asram datapaths before layout inference and tile-op lowering
+    mod = tilelang.transform.LegalizeSunmmioDataPath()(mod)
     mod = tilelang.transform.LayoutReducer()(mod)
     # Infer memory layouts — target-conditional
     if target_is_sunmmio(target):
